* Updated commons module inventory with new SHA256 hashes for headers. * Added platform-specific header guards for GLImpl and MetalImpl in the surface generation script. * Registered Python backends and integrated Python entry points for GPU recognition and tracking services. * Added Python backend registration for GPU recognition and tracking services. * Removed unused private header from Python implementation. * Updated public headers to include new Python entry point and backend registry.

+++ Application/src/ProcessedVideo/pv.cpp
@@ -1854,7 +1854,7 @@ Frame_t File::length() const {
 void File::set_average(const cv::Mat& average) {
     //tf::imshow("average", average);
     const auto required_channels = _header.encoding == meta_encoding_t::binary ? 1u : required_storage_channels(_header.encoding);
-    if(required_channels != average.channels()) /// just basic compatibility
+    if(int64_t(required_channels) != int64_t(average.channels())) /// just basic compatibility
     {
         throw InvalidArgumentException("Number of channels ",average.channels()," must match the encoding format ", _header.encoding," for the average image provided.");
     }
